requirements.txt
blis==0.7.9
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
cymem==2.0.7
idna==3.4
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.2
murmurhash==1.0.9
numpy==1.24.3
packaging==23.1
pathy==0.10.1
preshed==3.0.8
pydantic==1.10.7
pystache==0.6.0
PyYAML==6.0
requests==2.30.0
smart-open==6.3.0
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
thinc==8.1.10
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
urllib3==2.0.2
wasabi==1.1.1